add edit func

+++ tools/index.sh
@@ -2,13 +2,13 @@
 
 # 遍历posts目录下所有的html文件
 
-# Path conf
-#. ./tools/conf
-
 # generate new post
 date=`date '+%Y%m%d'`
 logfile=$PATH_LOG/$date"_index.log"
-rm $logfile
+if [ -e $logfile ]; then
+    rm $logfile
+fi
+
 $PHP $PATH_TOOL/php/index.php >> $logfile
 
 cat $logfile
