+++ tools/parse_md.sh
@@ -51,7 +51,10 @@ fi
 # parse md file
 date=`date '+%Y%m%d'`
 logfile=$PATH_LOG/$date"_debug.log"
-rm $logfile
+if [ -e $logfile ]; then
+    rm $logfile
+fi
+
 $PHP $PATH_TOOL/php/parse_md.php --file=$MD_FILENAME --all=$MD_ALL >> $logfile
 
 cat $logfile
